requirements.txt
Pillow
pytest
python-dotenv